feature: add pod resource watch
